refactor(hub): move blob strategy/factory layer — strategy → port/with/blob-strategy, withBlobs barrel → shape/via-blob (#629)
Partial move (re-scoped from the plan's whole-directory git mv): the
content-crypto machinery (blob-set/blob-compaction/export-blobs/mime-magic/
object-projection/import-external) STAYS in with-shape/blobs — its enclave
usage (chunk AEAD, hashing, per-blob CEK lifecycle) is not ViaCryptoCtx
material this phase and would otherwise trip via-enclave-isolation's EMPTY
allowlist. Only the declarative layer moves:

- with-shape/blobs/strategy.ts → port/with/blob-strategy.ts (the
  classified/i18n-strategy precedent); kernel spine specifiers updated and
  the 4 now-dead '../with-shape/blobs/strategy.js' port-layering
  grandfather entries retired.
- with-shape/blobs/{active,index}.ts → shape/via-blob/ (tsup blobs/index
  entry re-pointed; 12 test files re-pathed mechanically).
- with-shape/blobs joins SCHEMA_DECLARED_OR_INFRA_EXEMPT (its withBlobs()
  gate now lives in shape/via-blob, same bucket as with-party/sync).

Canary-verified via-enclave-isolation fires on a synthetic enclave import
from shape/via-blob (allowlist stays EMPTY). check:architecture green;
blob suites green unchanged.
