+++ packages/hub/__tests__/bundle-blobs-roundtrip.test.ts
@@ -22,7 +22,7 @@
 
 import { describe, it, expect } from 'vitest'
 import { createNoydb } from '../src/kernel/noydb.js'
-import { withBlobs } from '../src/with-shape/blobs/index.js'
+import { withBlobs } from '../src/shape/via-blob/index.js'
 import { withHistory } from '../src/with-commit/history/index.js'
 import {
   BLOB_INDEX_COLLECTION,
