+++ packages/hub/__tests__/export-blobs.test.ts
@@ -4,7 +4,7 @@
 import { describe, it, expect } from 'vitest'
 import type { NoydbStore, EncryptedEnvelope, VaultSnapshot } from '../src/index.js'
 import { ConflictError, ExportCapabilityError, createNoydb } from '../src/index.js'
-import { withBlobs } from '../src/with-shape/blobs/index.js'
+import { withBlobs } from '../src/shape/via-blob/index.js'
 import type { Noydb, Vault } from '../src/index.js'
 import { ExportBlobsAbortedError, EXPORT_AUDIT_COLLECTION } from '../src/with-shape/blobs/export-blobs.js'
 import { withTeam } from '../src/with-party/team/index.js'
