+++ packages/hub/src/kernel/collection-config.ts
@@ -33,7 +33,7 @@ import { NO_CRDT, type CrdtStrategy } from '../with-commit/crdt/strategy.js'
 import { NO_HISTORY, type HistoryStrategy } from '../with-commit/history/strategy.js'
 import { NO_I18N, type I18nStrategy } from '../port/with/i18n-strategy.js'
 import { NO_SYNC, type SyncStrategy } from '../with-party/team/sync-strategy.js'
-import { NO_BLOBS, type BlobStrategy } from '../with-shape/blobs/strategy.js'
+import { NO_BLOBS, type BlobStrategy } from '../port/with/blob-strategy.js'
 import { NO_AGGREGATE, type AggregateStrategy } from '../with-lookup/aggregate/strategy.js'
 import { NO_TIERS, type TiersStrategy } from '../with-audit/tiers/strategy.js'
 import { NO_SEARCH, type SearchStrategy } from '../with-lookup/search/strategy.js'
